requirements.txt
# Core Python packages for the entire project
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.5.1

# Data processing and database
kafka-python==2.0.2
psycopg2-binary==2.9.7
sqlalchemy==1.4.49
alembic==1.12.0
pandas==2.0.3
numpy==1.24.3

# HTTP and API
requests==2.31.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1
